Partial commit

+++ src/main/java/com/miro/assignment/dao/WidgetDao.java
@@ -7,8 +7,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface WidgetDao extends PagingAndSortingRepository<Widget, Integer> {
+public interface WidgetDao extends PagingAndSortingRepository<Widget, BigInteger> {
 
     @Query(value = "select w from Widget w where (w.x - w.width/2) >= 0 and (w.x + w.width/2) <= :upperX and (w.y - w.height/2) >= 0 and (w.y + w.height/2) <= :upperY")
     //@Query(value = "select w from Widget w")
