+++ src/main/java/com/miro/assignment/service/WidgetService.java
@@ -6,13 +6,14 @@ import com.miro.assignment.domain.WidgetRequest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 public interface WidgetService {
 
     Widget createWidget(WidgetRequest widget);
     Page<Widget> getAllWidgets(Pageable pageable, Optional<WidgetFilterCriteria> criteria);
-    Optional<Widget> getWidget(int id);
-    Optional<Widget> updateWidget(int id, WidgetRequest widget);
-    void deleteWidget(int id);
+    Optional<Widget> getWidget(BigInteger id);
+    Optional<Widget> updateWidget(BigInteger id, WidgetRequest widget);
+    void deleteWidget(BigInteger id);
 }
